Abstract the Bottom Navbar from Home.dart as separate widget
