+++ bicycle_trip_planner/lib/widgets/Home.dart
@@ -8,6 +8,8 @@ import 'package:bicycle_trip_planner/bloc/application_bloc.dart';
 import 'package:bicycle_trip_planner/widgets/Map.dart';
 import 'package:provider/provider.dart';
 
+import 'StationBar.dart';
+
 class Home extends StatefulWidget {
   const Home({Key? key}) : super(key: key);
 
@@ -16,55 +18,9 @@ class Home extends StatefulWidget {
 }
 
 class _HomeState extends State<Home> {
-
-  //********** Controllers **********
-
-  PageController stationsPageViewController = PageController();
-
-  //********** Station **********
-
-  void showExpandedList(List<Station> stations) {
-    showModalBottomSheet(
-        context: context,
-        builder: (BuildContext context) {
-          return Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.end,
-            children: <Widget>[
-              SizedBox(
-                  height: (56 * 6).toDouble(),
-                  child: Container(
-                      padding: const EdgeInsets.all(5),
-                      decoration: const BoxDecoration(
-                        color: Color(0xff345955),
-                      ),
-                      child: Column(
-                        children: [
-                          const Text(
-                            "Select a station",
-                            style: TextStyle(color: Colors.white),
-                          ),
-                          const SizedBox(height: 5),
-                          Expanded(
-                            child: ListView.builder(
-                                itemCount: stations.length,
-                                itemBuilder:
-                                    (BuildContext context, int index) =>
-                                    StationCard(index: index)),
-                          ),
-                        ],
-                      ))),
-            ],
-          );
-        });
-  }
-
-  //********** Widget **********
-
+  
   @override
   Widget build(BuildContext context) {
-
-    final applicationBloc = Provider.of<ApplicationBloc>(context);
 
     return Scaffold(
       body: SafeArea(
@@ -75,39 +31,7 @@ class _HomeState extends State<Home> {
             ],
           )
       ),
-      bottomNavigationBar: BottomAppBar(
-        elevation: 0,
-        color: const Color(0xff345955),
-        child: SizedBox(
-            height: 56.0,
-            child: Row(
-              children: [
-                IconButton(
-                  onPressed: () => showExpandedList(applicationBloc.stations),
-                  icon: const Icon(Icons.menu),
-                  color: Colors.white,
-                ),
-                Flexible(
-                  child: PageView.builder(
-                      controller: stationsPageViewController,
-                      physics: const PageScrollPhysics(),
-                      scrollDirection: Axis.horizontal,
-                      itemCount: applicationBloc.stations.length,
-                      itemBuilder: (BuildContext context, int index) =>
-                          StationCard(index: index)),
-                ),
-                SizedBox(
-                  width: 30.0,
-                  child: IconButton(
-                    padding: const EdgeInsets.all(0),
-                    onPressed: () => stationsPageViewController.jumpTo(0),
-                    icon: const Icon(Icons.first_page),
-                    color: Colors.white,
-                  ),
-                ),
-              ],
-            )),
-      ),
+      bottomNavigationBar: StationBar(),
     );
   }
